Added FCC C++ DLL Project
Added a C++ DLL project to the solution. This DLL is meant to contain a
function that takes a string of bytes and a string stating what kind of
data is being input. It returns a compressed JavaScript, CSS or HTML
string. In the future it will also have functions for any combination of
the three. None of these function have been implemented yet, there
containing file has just been inserted. Make sure to read the Readme if
you want to use this project.